+++ resources/task.csv
@@ -2,8 +2,7 @@ id,type,name,status,description,epic, startTime, duration
 1,TASK,Помыть посуду,NEW,помыть посуду горячей водой,0,2023-12-12T20:00,40,
 5,TASK,Помыть посуду2,NEW,помыть посуду горячей водой2,0,2023-12-12T14:00,40,
 2,EPIC,Переезд,NEW,Переехать в другую квартиру,0,2023-12-12T15:00,85,
-6,EPIC,Переезд3,NEW,Переехать в другую квартиру3,0,2024-04-12T15:49:57.445671200,0,
+6,EPIC,Переезд3,NEW,Переехать в другую квартиру3,0,2024-04-12T16:31:14.960297900,0,
 3,SUBTASK,Собрать коробки,NEW,собрать в коробки вещи,2,2023-12-12T15:00,30,
 4,SUBTASK,Собрать коробки2,NEW,собрать в коробки вещи2,2,2023-12-12T16:00,25,
 
-1
